+++ PSServicePrincipal/internal/functions/Connect-ToAzureInteractively.ps1
@@ -22,52 +22,65 @@
     # $ApplicationID = 'xxxxxxx-xxxx-xxxx-xxxx-xxxxxxxxxxxx'
     # $CertificateThumbprint = 'xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx'
 
-    try
+    if(-NOT $script:AdSessionFound)
     {
-        if(-NOT $script:adSessionFound)
+        if($PSVersionTable.PSEdition -eq "Core")
         {
-            if($PSVersionTable.PSEdition -eq "Core")
-            {
-                Write-PSFMessage -Level Host -Message "At this time AzureAD PowerShell module does not work on PowerShell Core. Please use PowerShell version 5 or 6." -Once "PS Core Doesn't work with AzureAD" -FunctionName "Connect-ToAzureInteractively"
-                $script:runningOnCore = $true
-            }
-            elseif($PSVersionTable.PSEdition -eq "Desktop")
-            {
-
-
-                try
-                {
-                    Connect-AzureAD -TenantId $TenantID -ApplicationId $ApplicationID -CertificateThumbprint $CertificateThumbprint -ErrorAction Stop
-                    Write-PSFMessage -Level Host -Message "Connected to AzureAD with interactive logon"  -Once "Interactively Connected to AzureAD" -FunctionName "Connect-ToAzureInteractively"
-                    return
-                }
-                catch
-                {
-                    Write-PSFMessage -Level Host -Message "Interactive logon to AzureAD failed. Defaulting to default connection state" -Once "Interactive Logon Failed" -FunctionName "Connect-ToAzureInteractively"
-                    Connect-ToAzureDefaultConnection
-                }
-            }
+            Write-PSFMessage -Level Host -Message "At this time AzureAD PowerShell module does not work on PowerShell Core. Please use PowerShell version 5 or 6." -Once "PS Core Doesn't work with AzureAD" -FunctionName "Connect-ToAzureInteractively"
+            $script:runningOnCore = $true
         }
-
-        if(-NOT $script:azSessionFound)
+        elseif($PSVersionTable.PSEdition -eq "Desktop")
         {
             try
             {
-                Connect-AzAccount -TenantId $TenantID -ApplicationId $ApplicationID -CertificateThumbprint $CertificateThumbprint -ErrorAction Stop
-                Write-PSFMessage -Level Host -Message "Connected to AzureAZ with interactive logon" -Once "Interactively Connected to AzureAZ" -FunctionName "Connect-ToAzureInteractively"
+                Connect-AzureAD -TenantId $TenantID -ApplicationId $ApplicationID -CertificateThumbprint $CertificateThumbprint -ErrorAction Stop
+                Write-PSFMessage -Level Host -Message "Connected to AzureAD with interactive logon"  -Once "Interactively Connected to AzureAD" -FunctionName "Connect-ToAzureInteractively"
+                $script:AdSessionFound = $true
             }
             catch
             {
-                Write-PSFMessage -Level Host -Message "Interactive logon to AzureAZ failed. Defaulting to default connection state" -Once "Interactive Connection Failed" -FunctionName "Connect-ToAzureInteractively"
-                Connect-ToAzureDefaultConnection
+                Write-PSFMessage -Level Host -Message "Interactive logon to AzureAD failed. Defaulting to default connection state" -Once "Interactive Logon Failed" -FunctionName "Connect-ToAzureInteractively"
+                $script:AdSessionFound = $false
+
+                try
+                {
+                    $Credentials = Get-Credential
+                    $script:AzSessionInfo = Connect-AzAccount -Credential $Credentials -ErrorAction Stop
+                    $script:AzSessionFound = $true
+                    Write-PSFMessage -Level Host -Message "Interactive logon to AzureAD failed. Defaulting to default connection state" -Once "Interactive Logon Failed" -FunctionName "Connect-ToAzureInteractively"
+                }
+                catch
+                {
+                    Stop-PSFFunction -Message $_ -Cmdlet $PSCmdlet -ErrorRecord $_ -EnableException $true
+                }
             }
         }
-
     }
-    catch
+
+    if(-NOT $script:AzSessionFound)
     {
-        Write-PSFMessage -Level Host -Message "Interactive connection failed. Defaulting to manual connection" -Once "Interactive Connection Failed. Start Manual Connection" -FunctionName "Connect-ToAzureInteractively"
-        Connect-ToAzureDefaultConnection -AzSessionFound $azSessionFound
+        try
+        {
+            Connect-AzAccount -TenantId $TenantID -ApplicationId $ApplicationID -CertificateThumbprint $CertificateThumbprint -ErrorAction Stop
+            Write-PSFMessage -Level Host -Message "Connected to AzureAZ with interactive logon" -Once "Interactively Connected to AzureAZ" -FunctionName "Connect-ToAzureInteractively"
+            $script:AzSessionFound = $true
+        }
+        catch
+        {
+            Write-PSFMessage -Level Host -Message "Interactive logon to AzureAZ failed. Defaulting to default connection state" -Once "Interactive Connection Failed" -FunctionName "Connect-ToAzureInteractively"
+            $script:AzSessionFound = $false
+
+            try
+            {
+                $Credentials = Get-Credential
+                $script:AzSessionInfo = Connect-AzAccount -Credential $Credentials -ErrorAction Stop
+                $script:AzSessionFound = $true
+            }
+            catch
+            {
+                Stop-PSFFunction -Message $_ -Cmdlet $PSCmdlet -ErrorRecord $_ -EnableException $true
+            }
+        }
     }
 }