+++ PSServicePrincipal/changelog.md
@@ -1,6 +1,5 @@
 ﻿# Changelog
 ## 1.0.5 (2020-05-05)
- - New: Modifications to the logging subsystem
  - New: Asked for requested feature - Create Enterprise or tenant owned application
- - Upd: Documentation, incremented module version
- - Fix: Mac folder redirection for log folder path
+ - Upd: Documentation, incremented module version. Added checks for PowerShell core as AzureAD will not work on PowerShell core.
+ - Fix: Mac folder redirection for log folder path.