+++ PSServicePrincipal/changelog.md
@@ -1,5 +1,6 @@
 ﻿# Changelog
 ## 1.0.5 (2020-05-05)
- - New: Asked for requested feature - Create Enterprise or tenant owned application
+ - New: Ability to create Enterprise and or tenant owned applications with service principals
+ - New: Ability to create self-signed certificates to be used for (CBA) Certificate Based Authentication
  - Upd: Documentation, incremented module version. Added checks for PowerShell core as AzureAD will not work on PowerShell core.
- - Fix: Mac folder redirection for log folder path.
+ - Fix: File save for New-SelfSignedCert - Test for file path and ask to create it if path is not found