+++ PSServicePrincipal/changelog.md
@@ -1,5 +1,5 @@
 ﻿# Changelog
-## 1.0.7 (2020-05-05)
+## 1.0.9 (2020-05-12)
  - New: Ability to create Enterprise and or tenant owned applications with service principals
  - New: Ability to create self-signed certificates to be used for (CBA) Certificate Based Authentication
  - Upd: Documentation, incremented module version. Added checks for PowerShell core as AzureAD will not work on PowerShell core.